refactor: API 요청 로직 모듈 분리 및 코드 정리
